winbuild now also may precompile targets
ascii targets can be removed for this purpose, target-compiler was
changed to fit to winbuild. Program could work with mhbin only, but that
still would result in the problem that the repo will not find them.
target-compiler shall not delete compressedtargets.npz in system or
contarget folder, if no targets are existent, since that would destroy
the only available npz targets as well!

+++ core/importfiles.py
@@ -382,7 +382,7 @@ class TargetASCII():
                     content[name[:-7]] = arr
         return content
 
-    def compressAllTargets(self, sourcefolder, destfile, verbose=0):
+    def compressAllTargets(self, sourcefolder, destfile, verbose=0, remove=True):
         content = self.loadAllTargets(sourcefolder, verbose)
         howmany = len(content)
         if howmany > 0:
@@ -392,7 +392,8 @@ class TargetASCII():
         else:
             if verbose > 0:
                 print ("No content for: " + destfile)
-            if os.path.exists(destfile):
-                os.remove(destfile)
+            if remove:
+                if os.path.exists(destfile):
+                    os.remove(destfile)
         return howmany
